requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.1.6
astroid==3.0.3
attrs==24.2.0
bcrypt==4.2.1
black==23.11.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
commitizen==3.12.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
decli==0.6.2
deepgram-sdk
dill==0.3.9
distro==1.9.0
dotty-dict==1.3.1
ecdsa==0.19.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.19.0
gradio_client==0.10.0
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
isort==5.13.2
Jinja2==3.1.4
jiter==0.8.0
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langchain-together==0.2.0
langsmith==0.1.147
lxml-stubs==0.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.3
mccabe==0.7.0
mdurl==0.1.2
motor
mpmath==1.3.0
multidict==6.1.0
mypy==1.7.1
mypy-extensions==1.0.0
narwhals==1.15.1
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.55.3
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241126
passlib==1.7.4
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.3.6
prompt-toolkit==3.0.36
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pylint==3.0.2
pymongo
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-gitlab==3.15.0
python-jose==3.3.0
python-multipart==0.0.6
python-semantic-release==8.3.0
pytz==2024.2
PyYAML==6.0.2
questionary==2.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
ruff==0.8.1
safehttpx==0.1.1
safetensors==0.4.5
semantic-version==2.10.0
shellingham==1.5.4
silero==0.4.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.35
srt==3.5.3
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typer==0.14.0
types-beautifulsoup4==4.12.0.20241020
types-cffi==1.16.0.20240331
types-colorama==0.4.15.20240311
types-docutils==0.21.0.20241128
types-html5lib==1.1.11.20241018
types-jsonschema==4.23.0.20240813
types-passlib==1.7.7.20240819
types-Pillow==10.2.0.20240822
types-protobuf==5.28.3.20241030
types-psutil==6.1.0.20241102
types-pyasn1==0.6.0.20240913
types-Pygments==2.18.0.20240506
types-pyOpenSSL==24.1.0.20240722
types-python-jose==3.3.4.20240106
types-pytz==2024.2.0.20241003
types-PyYAML==6.0.12.20240917
types-redis==4.6.0.20241004
types-regex==2024.11.6.20241108
types-requests==2.32.0.20241016
types-setuptools==75.6.0.20241126
types-tabulate==0.9.0.20240106
types-tqdm==4.67.0.20241119
types-ujson==5.10.0.20240515
types-xmltodict==0.14.0.20241009
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.24.0.post1
vosk==0.3.45
wcwidth==0.2.13
websockets==11.0.3
yarl==1.18.3
zipp==3.21.0

